added a test
added a basic test to check if numpy is installed and if np.pi is indeed a float. Currently, the system fails and gives an error if numpy isn't installed, and passes if it is. This isn't quite the T/F behavior I think is typical of testing suites, but I think it's fine for me, for now.